refactoring - Erigon in services directory
